SpellingCheck_Natas/requirements.txt
absl-py==0.9.0
args==0.1.0
beautifulsoup4==4.8.2
blis==0.4.1
boto==2.49.0
boto3==1.11.9
botocore==1.14.9
bs4==0.0.1
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
clint==0.5.1
ConfigArgParse==1.0
cycler==0.10.0
cymem==2.0.3
Distance==0.1.3
docutils==0.15.2
Flask==1.1.1
future==0.18.2
gensim==3.8.1
google-auth==1.11.0
google-auth-oauthlib==0.4.1
grpcio==1.26.0
idna==2.8
importlib-metadata==1.4.0
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mikatools==0.0.7
murmurhash==1.0.2
natas==1.0.4
nltk==3.4.5
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.1.2.30
OpenNMT-py==1.0.0
pdf2image==1.11.0
Pillow==7.0.0
plac==1.1.3
preshed==3.0.2
protobuf==3.11.2
pt-core-news-sm==2.2.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyonmttok==1.18.1
pyparsing==2.4.6
pytesseract==0.3.2
python-dateutil==2.8.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.2
scikit-learn==0.22.1
scipy==1.4.1
sentencepiece==0.1.85
six==1.14.0
sklearn==0.0
smart-open==1.9.0
soupsieve==1.9.5
spacy==2.2.3
srsly==1.0.1
tensorboard==2.1.0
thinc==7.3.1
torch==1.0.0
torchtext==0.5.0
tqdm==4.30.0
urllib3==1.25.8
wasabi==0.6.0
Werkzeug==0.16.1
zipp==2.1.0
